* shown when teamsize is equal
