+++ src/components/TeamSizePicker/TeamSizes.js
@@ -5,6 +5,8 @@ import Grid from '@material-ui/core/Grid';
 import Typography from '@material-ui/core/Typography';
 import { withStyles } from '@material-ui/core/styles';
 import { Divider, List, ListItem, ListItemText } from '@material-ui/core';
+import shuffle from '../../utils/shuffle';
+import chunkArray from '../../utils/chunckArray';
 
 const styles = {
   card: {
@@ -24,39 +26,8 @@ const styles = {
   }
 };
 
-const createTeams = (members, size) => {
-  return chunkArray(members, size);
-};
-
-function chunkArray(arr, chunkCount) {
-  const chunks = [];
-  while (arr.length) {
-    const chunkSize = Math.ceil(arr.length / chunkCount--);
-    const chunk = arr.slice(0, chunkSize);
-    chunks.push(chunk);
-    arr = arr.slice(chunkSize);
-  }
-  return chunks;
-}
-
-function shuffle(array) {
-  let currentIndex = array.length,
-    temporaryValue,
-    randomIndex;
-
-  while (0 !== currentIndex) {
-    randomIndex = Math.floor(Math.random() * currentIndex);
-    currentIndex -= 1;
-    temporaryValue = array[currentIndex];
-    array[currentIndex] = array[randomIndex];
-    array[randomIndex] = temporaryValue;
-  }
-
-  return array;
-}
-
 const TeamSizes = ({ size, members, classes }) => {
-  const teams = createTeams(shuffle(members), size);
+  const teams = chunkArray(shuffle(members), size);
   return (
     <Grid container direction="row" justify="center" alignItems="center" spacing={24}>
       {teams.map((team, index) => (
