+++ src/utils/chunckArray.js
@@ -1,0 +1,10 @@
+export default (arr, chunkCount) => {
+  const chunks = [];
+  while (arr.length) {
+    const chunkSize = Math.ceil(arr.length / chunkCount--);
+    const chunk = arr.slice(0, chunkSize);
+    chunks.push(chunk);
+    arr = arr.slice(chunkSize);
+  }
+  return chunks;
+};
